+++ app/src/main/java/ru/netology/nmedia/dto/Post.kt
@@ -1,4 +1,4 @@
-package ru.netology.nmedia.post
+package ru.netology.nmedia.dto
 
 data class Post (
     val id:Long,
