+++ app/src/main/res/layout/activity_main.xml
@@ -4,59 +4,55 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".activity.MainActivity">
 
     <ImageView
         android:id="@+id/avatar"
         android:layout_width="48dp"
         android:layout_height="48dp"
-        android:layout_marginTop="16dp"
         android:layout_marginStart="16dp"
+        android:layout_marginTop="16dp"
         android:layout_marginBottom="16dp"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintBottom_toBottomOf="@id/topBarrier"
-        tools:srcCompat="@sample/posts.json/data/authorAvatar"
         android:src="@mipmap/post_avatar_drawable"
-        />
+        app:layout_constraintBottom_toBottomOf="@id/topBarrier"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        tools:srcCompat="@sample/posts.json/data/authorAvatar" />
 
     <TextView
         android:id="@+id/author"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        app:layout_constraintBottom_toTopOf="@id/published"
-        app:layout_constraintTop_toTopOf="@+id/avatar"
-        app:layout_constraintStart_toEndOf="@id/avatar"
-        app:layout_constraintEnd_toStartOf="@id/menu"
         android:layout_marginStart="16dp"
-        tools:text="@sample/posts.json/data/author"
-        android:maxLines="1"
         android:ellipsize="end"
-        />
+        android:maxLines="1"
+        app:layout_constraintBottom_toTopOf="@id/published"
+        app:layout_constraintEnd_toStartOf="@id/menu"
+        app:layout_constraintStart_toEndOf="@id/avatar"
+        app:layout_constraintTop_toTopOf="@+id/avatar"
+        tools:text="@sample/posts.json/data/author" />
 
     <TextView
         android:id="@+id/published"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        app:layout_constraintBottom_toBottomOf="@id/avatar"
-        app:layout_constraintTop_toBottomOf="@id/author"
-        app:layout_constraintStart_toEndOf="@id/avatar"
         android:layout_marginStart="16dp"
-        tools:text="@sample/posts.json/data/published"
-        />
+        app:layout_constraintBottom_toBottomOf="@id/avatar"
+        app:layout_constraintStart_toEndOf="@id/avatar"
+        app:layout_constraintTop_toBottomOf="@id/author"
+        tools:text="@sample/posts.json/data/published" />
 
     <ImageButton
         android:id="@+id/menu"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:layout_marginEnd="16dp"
-        tools:srcCompat="@drawable/icon_menu"
         android:layout_marginTop="16dp"
+        android:layout_marginEnd="16dp"
         android:layout_marginRight="16dp"
         android:background="@android:color/transparent"
-        />
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        tools:srcCompat="@drawable/icon_menu" />
 
     <androidx.constraintlayout.widget.Barrier
         android:id="@+id/topBarrier"
@@ -69,16 +65,15 @@
         android:id="@+id/content"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        app:layout_constraintTop_toBottomOf="@id/topBarrier"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:layout_marginTop="16dp"
-        android:layout_marginBottom="16dp"
         android:layout_marginStart="16dp"
+        android:layout_marginTop="16dp"
         android:layout_marginEnd="16dp"
-        tools:text="@sample/posts.json/data/content"
+        android:layout_marginBottom="16dp"
         android:autoLink="web"
-        />
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/topBarrier"
+        tools:text="@sample/posts.json/data/content" />
 
     <androidx.constraintlayout.widget.Barrier
         android:id="@+id/contentBarrier"
@@ -91,12 +86,11 @@
         android:id="@+id/like"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        app:layout_constraintTop_toTopOf="@id/contentBarrier"
-        app:layout_constraintStart_toStartOf="parent"
-        android:src="@drawable/icon_like"
         android:layout_marginStart="16dp"
         android:layout_marginTop="16dp"
-        />
+        android:src="@drawable/icon_like"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="@id/contentBarrier" />
 
     <TextView
         android:id="@+id/countLike"
@@ -113,12 +107,11 @@
         android:id="@+id/share"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        app:layout_constraintTop_toTopOf="@id/contentBarrier"
-        app:layout_constraintStart_toEndOf="@id/countLike"
-        android:src="@drawable/icon_share"
         android:layout_marginStart="16dp"
         android:layout_marginTop="16dp"
-        />
+        android:src="@drawable/icon_share"
+        app:layout_constraintStart_toEndOf="@id/countLike"
+        app:layout_constraintTop_toTopOf="@id/contentBarrier" />
 
     <TextView
         android:id="@+id/countShare"
@@ -133,23 +126,23 @@
 
     <ImageView
         android:id="@+id/views"
-        android:src="@drawable/icon_views"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        app:layout_constraintEnd_toStartOf="@id/countViews"
-        app:layout_constraintTop_toTopOf="@id/contentBarrier"
-        android:layout_marginEnd="16dp"
         android:layout_marginTop="16dp"
-        />
+        android:layout_marginEnd="16dp"
+        android:src="@drawable/icon_views"
+        app:layout_constraintEnd_toStartOf="@id/countViews"
+        app:layout_constraintTop_toTopOf="@id/contentBarrier" />
+
     <TextView
         android:id="@+id/countViews"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="16dp"
         android:layout_marginTop="16dp"
+        android:layout_marginEnd="16dp"
         android:text="5"
         app:layout_constraintEnd_toEndOf="parent"
-        android:layout_marginEnd="16dp"
-        app:layout_constraintTop_toTopOf="@id/contentBarrier"/>
+        app:layout_constraintTop_toTopOf="@id/contentBarrier" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
